requirements.txt
fastapi
uvicorn
motor
python-dotenv
PyJWT
aiosmtplib
pydantic
email-validator